Solution/pyproject.toml
[tool.poetry]
name = "Solution"
version = "0.1.0"
description = ""
authors = ["drinkwaterknow <a.kashipov@rt-solar.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
cryptography = "^35.0.0"
python_binance = "^1.0.15"
poetry = "^1.1.11"
ftx = "^1.0.2"
ciso8601 = "^2.2.0"
isort = "^5.9.3"
autopep8 = "^1.6.0"
pyTelegramBotAPI = "^4.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
